+++ python/packaging/client/setup.py
@@ -169,6 +169,7 @@ try:
         "pyspark.sql.connect.shell",
         "pyspark.sql.connect.streaming",
         "pyspark.sql.connect.streaming.worker",
+        "pyspark.sql.eval_handlers",
         "pyspark.sql.functions",
         "pyspark.sql.pandas",
         "pyspark.sql.plot",
